chore: extract the code to evaluate the attributes of tag into a module
